refactor: split provider metadata and add routing tests
- move metadata, routing and error helpers out of extension.ts

- add a timeout to models.dev metadata refresh and cover parser/routing cases

- document corrected fallback limits in the changelog
